Docs new apis (#153)
* Add Anthropics and Groq API configurations

Created new API configurations for Anthropic and Groq in the examples section. These configurations will initialize the given APIs with selected LLM models. Also, added code to validate if the respective API keys are set in the environment variables.

* Added a Jupyter Notebook(Anthropics, Groq)

Added Anthropic API Documentation: Added a Jupyter Notebook detailing the usage and configuration of the Anthropics API to the docs folder.
Added Groq API Documentation: Added a Jupyter Notebook that covers the usage of the Groq API to the docs folder.

* Update class names and add dependencies

Refactored 'AnthropicApiLLM' and 'GroqApiLLM' class names to 'LLM' in AnthropocApi.py and GroqApi.py files. Also, the 'llama-index-llms-anthropic' and 'llama-index-llms-groq' dependencies were added to the pyproject.toml file.

---------

Co-authored-by: 김윤웅_DEV_SKCC <yunwoong@sk.com>